Fix wrong data type in CpuShare computation

+++ pylibs/plancton/__init__.py
@@ -214,7 +214,7 @@ class Plancton(Daemon):
     c = { "Cmd"        : self.conf["docker_cmd"],
           "Image"      : self.conf["docker_image"],
           "Hostname"   : "plancton-%s-%s" % (self._hostname, uuid),
-          "HostConfig" : { "CpuShares"   : self.conf["cpus_per_dock"]*1024/cpu_count(),
+          "HostConfig" : { "CpuShares"   : int(self.conf["cpus_per_dock"]*1024/cpu_count()),
                            "NetworkMode" : "bridge",
                            "SecurityOpt" : ["apparmor:docker-allow-ptrace"] if apparmor_enabled() else [],
                            "Binds"       : [ x+":ro,Z" for x in self.conf["binds"] ],
